Add files via upload
Submission for Fork the World design bounty.  The circle in the head of the octopus is a masked layer, so can be used as a template for NFT achievements, Player portrait frames, whatever your heart desires.  

I have included all format's which might be pertinent.  If I have forgotten any please let me know and I'll get them to you.